# docs/collections/api-specification.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/invitation/invite-to-team": {
      "post": {
        "operationId": "InvitationController_inviteToTeam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamInviteInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "invitation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invitation/invite-to-project": {
      "post": {
        "operationId": "InvitationController_inviteToProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInviteInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "invitation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invitation/{invitationId}/accept": {
      "patch": {
        "operationId": "InvitationController_acceptInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "invitation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "The Secret Store api",
    "description": "This is the api for the secret store project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "name": "JWT",
        "type": "http",
        "description": "Some routes require a JWT token"
      }
    },
    "schemas": {
...
      "CreateTeamInviteInputDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProjectInviteInputDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}