# libs/api-interfaces/src/dtos/invitation/create-project-invite.input.dto.ts
import { ProjectAccessLevel } from '../../enums';
import { IsEmail, IsEnum, IsMongoId, IsNotEmpty } from 'class-validator';

export class CreateProjectInviteInputDto {
  @IsEmail()
  email: string;

  @IsMongoId()
  @IsNotEmpty()
  projectId: string;

  @IsEnum(ProjectAccessLevel)
  accessLevel: ProjectAccessLevel = ProjectAccessLevel.MEMBER;
}