# libs/api-interfaces/src/dtos/invitation/create-team-invite.input.dto.ts
import { Privilege } from '../../enums';
import { IsEmail, IsEnum, IsMongoId, IsNotEmpty } from 'class-validator';

export class CreateTeamInviteInputDto {
  @IsEmail()
  email: string;

  @IsMongoId()
  @IsNotEmpty()
  teamId: string;

  @IsEnum(Privilege)
  privilege: Privilege = Privilege.viewer;
}